+++ engine/examples/audioTest/resources/levels/audioTest/audioTestController.js
@@ -7,32 +7,32 @@ function controller(api) {
     const defaultColor = "white";
 
     const soundtrackList = api.world.getElementsFromLayer(2);
-    soundtrackList[hoveredSoundtrack].settings.color = selectColor;
+    soundtrackList[hoveredSoundtrack].settings.settings.color = selectColor;
 
     api.keyboard.pressEvent(e => {
         if(hoveredSoundtrack > 0) {
             hoveredSoundtrack--;
-            soundtrackList[hoveredSoundtrack+1].settings.color = defaultColor;
+            soundtrackList[hoveredSoundtrack+1].settings.settings.color = defaultColor;
         }
 
         else {
             hoveredSoundtrack = soundtrackList.length-1;
-            soundtrackList[0].settings.color = defaultColor;
+            soundtrackList[0].settings.settings.color = defaultColor;
         }
-        soundtrackList[hoveredSoundtrack].settings.color = selectColor;
+        soundtrackList[hoveredSoundtrack].settings.settings.color = selectColor;
     }, "arrLeft");
 
     api.keyboard.pressEvent(e => {
         if(hoveredSoundtrack < soundtrackList.length-1) {
             hoveredSoundtrack++;
-            soundtrackList[hoveredSoundtrack-1].settings.color = defaultColor;
+            soundtrackList[hoveredSoundtrack-1].settings.settings.color = defaultColor;
         }
 
         else {
             hoveredSoundtrack = 0;
-            soundtrackList[soundtrackList.length-1].settings.color = defaultColor;
+            soundtrackList[soundtrackList.length-1].settings.settings.color = defaultColor;
         }
-        soundtrackList[hoveredSoundtrack].settings.color = selectColor;
+        soundtrackList[hoveredSoundtrack].settings.settings.color = selectColor;
     }, "arrRight");
 
     api.keyboard.pressEvent(e => {