+++ project-editor/src/renderer/assets/js/canvas.js
@@ -1,10 +1,12 @@
-import RPGinia from '../../../../../engine/src/RPGinia';
+import { ipcRenderer } from 'electron';
+
+import RPGinia from '../../../../../engine/src/RPGinia'; // Import engine
 
 const appPath = process.env.NODE_ENV === 'development'
                 ? `http://localhost:9080`
                 : `file://${__dirname}/`;
 
-let engine, app, world, kb, cam, loop;
+let engine, app, world, cam, loop;
 let store, storeGetters;
 
 export default function initPlayground(data, projStore) {
@@ -18,12 +20,11 @@ export default function initPlayground(data, projStore) {
         ]
     );
     world = new app.World();
-    kb = new app.Keyboard();
     cam = new app.Camera();
 
     store = projStore;
     storeGetters = store.getters;
-    
+
     // On app resizing
     document.body.onresize = () => {
         app.sizes = [
@@ -32,51 +33,52 @@ export default function initPlayground(data, projStore) {
         ];
     }
 
-    kb.addKey('arrUp', 38);
-    kb.addKey('arrDown', 40);
-    kb.addKey('arrLeft', 37);
-    kb.addKey('arrRight', 39);
-    kb.addKey('alt', 18);
-
+    // Initialize the playground's world
     world.initialize({
         app: app,
         camera: cam,
         levels: data
     });
 
-    function drawBorders() {
-        console.log(app.context)
-    }
-    drawBorders();
-
-    // If alt key and left mouse button are pressed - move camera
     app.canvas.onmousemove = e => { 
+        // If alt key and left mouse button are pressed - move camera
         if(e.altKey && e.buttons === 1) cam.move(e.movementX, e.movementY);
     }
 
-    // Select object in canvas
+    // Select object in playground
     app.canvas.onclick = e => {
         const objectList = storeGetters.projectObjects;
         const selectedObjects = storeGetters.selectedObjects;
 
-        for(let i in objectList) {
-            if(
-                !e.altKey &&
-                
-                e.offsetX >= objectList[i].settings.coords[0] + cam.x &&
-                e.offsetX <= objectList[i].settings.coords[0] + objectList[i].settings.coords[2] + cam.x &&
-                
-                e.offsetY >= objectList[i].settings.coords[1] + cam.y &&
-                e.offsetY <= objectList[i].settings.coords[1] + objectList[i].settings.coords[3] + cam.y
-            ) {
-                if(selectedObjects.indexOf(objectList[i].$id) === -1) {
-                    for(let j in selectedObjects) {
-                        store.dispatch('unselectObject', {
-                            from: j, 
-                            to: 1
-                        });
+        if(!e.altKey) {
+            for(let i in objectList) {
+                if(
+                    e.offsetX >= objectList[i].settings.coords[0] + cam.x &&
+                    e.offsetX <= objectList[i].settings.coords[0] + objectList[i].settings.coords[2] + cam.x &&
+                    
+                    e.offsetY >= objectList[i].settings.coords[1] + cam.y &&
+                    e.offsetY <= objectList[i].settings.coords[1] + objectList[i].settings.coords[3] + cam.y
+                ) {
+                    if(selectedObjects.indexOf(objectList[i].$id) === -1) {
+                        for(let j in selectedObjects) {
+                            store.dispatch('unselectObject', {
+                                from: j,
+                                to: 1
+                            });
+                        }
+                        
+                        store.dispatch('selectObject', objectList[i].$id);
+                        return;
                     }
-                    store.dispatch('selectObject', objectList[i].$id);
+
+                    else {
+                        for(let j in selectedObjects) {
+                            store.dispatch('unselectObject', {
+                                from: j, 
+                                to: 1
+                            });
+                        }
+                    }
                 }
 
                 else {
@@ -85,26 +87,50 @@ export default function initPlayground(data, projStore) {
                             from: j, 
                             to: 1
                         });
+                        return;
                     }
-                }
-                return;
-            }
-
-            else {
-                for(let j in selectedObjects) {
-                    store.dispatch('unselectObject', {
-                        from: j, 
-                        to: 1
-                    });
                 }
             }
         }
     }
 
+    // Create object
+    ipcRenderer.on('createObject', (e, object) => {
+        let obj = world.createElement(object);
+        store.dispatch('addObject', world.getElementByName(obj.name));
+    });
+
+    // Delete object
+    ipcRenderer.on('deleteObject', (e, objectIndex) => { store.dispatch('deleteObject', objectIndex) });
+
     loop = () => {
         app.clearPlayground();
 
         world.draw();
+
+        for(let selectedId of storeGetters.selectedObjects) {
+            const selectedObject = storeGetters.projectObjects[storeGetters.projectObjects.findIndex(item => item.$id === selectedId)];
+
+            // -------
+            //  Lines
+            // -------
+            app.context.strokeStyle = '#ffffff';
+            app.context.lineWidth = 2;
+            app.context.setLineDash([5, 5]);
+            app.context.beginPath();
+            
+            app.context.rect(
+                cam.x + selectedObject.settings.coords[0],
+                cam.y + selectedObject.settings.coords[1],
+                selectedObject.settings.coords[2],
+                selectedObject.settings.coords[3]
+            );
+
+            app.context.stroke();
+            app.context.setLineDash([]);
+            app.context.lineWidth = 1;
+            app.context.strokeStyle = '#000000';
+        }
 
         requestAnimationFrame(loop);
     }