+++ project-editor/src/renderer/assets/js/clipboard.js
@@ -1,0 +1,7 @@
+export function copy(window) {
+    window.webContents.send('copySelectedObjects');
+}
+
+export function paste(window) {
+    window.webContents.send('pasteSelectedObjects');
+}