+++ project-editor/src/renderer/components/EditorComponents/Blocks/BlockTitle.vue
@@ -5,8 +5,8 @@
                 <p>{{ title }}</p>
             </div>
             <div class="right">
-                <button class="btn_transparent" id="add" v-if="sideButtons || addButton">+</button>
-                <button class="btn_transparent" id="delete" v-if="sideButtons || deleteButton">-</button>
+                <button class="btn_transparent" id="add" v-if="sideButtons || addButton" @click="$emit('add')">+</button>
+                <button class="btn_transparent" id="delete" v-if="sideButtons || deleteButton" @click="$emit('delete')">-</button>
             </div>
         </div> 
         <p v-else>{{ title }}</p>