+++ project-editor/src/renderer/components/EditorComponents/Playground.vue
@@ -1,0 +1,5 @@
+<template>
+    <div class="canvas_container">
+        <canvas id="playground"></canvas>
+    </div>
+</template>