+++ project-editor/src/renderer/components/ProjectExistingError.vue
@@ -30,7 +30,7 @@ export default {
         },
 
         openProject: function() {
-            ipcRenderer.send('openProject')
+            ipcRenderer.send('openProject');
         }
     }
 }